backend/requirements.txt
# requirements.txt
# Lists Python dependencies for backend
# Example: flask, flask-cors, requests, python-dotenv


flask
flask-cors
requests
python-dotenv
